Course crud operations

+++ backend/crud/course_crud.py
@@ -1,4 +1,5 @@
 import json
+from http.client import HTTPException
 from uuid import UUID
 
 from ..database.db import supabase
@@ -6,6 +7,11 @@ from ..routers.req.courses_req import CreateCourseReq, RegisterUserReq, UpdateCo
 from fastapi.encoders import jsonable_encoder
 
 courses_table = supabase.table('courses')
+
+class UserNotEnrolledException(Exception):
+    pass
+class NoPermissionException(Exception):
+    pass
 
 def register_course(req: RegisterUserReq):
     try:
@@ -17,34 +23,57 @@ def register_course(req: RegisterUserReq):
     except Exception as e:
         return None
 
-def get_courses():
+def get_courses(user_id: str):
     try:
-        response = courses_table.select("*").execute()
+        user_course_ids_res = supabase.table("user_courses").select("course_id").eq('user_id', user_id).execute()
+        user_courses_dict = user_course_ids_res.data
+        user_courses_list = list(map(lambda n: n['course_id'], user_courses_dict))
+        response = courses_table.select("*").in_("id", user_courses_list).execute()
         return response
     except Exception as e:
         return None
 
-def get_course_by_id(c_id: int):
+def get_course_by_id(c_id: int, user_id: str):
     try:
-        return courses_table.select("*").eq('id', c_id).execute()
-    except Exception as e:
+        user_course = (supabase.from_("user_courses").select("*").eq("course_id", c_id)
+                       .eq("user_id", user_id).execute())
+        if len(user_course.data) > 0:
+            return courses_table.select("*").eq('id', c_id).execute()
+        raise UserNotEnrolledException
+    except ValueError as e:
         return None
 
-def update_course(req: UpdateCourseReq):
+def update_course(req: UpdateCourseReq, user_id: str, admin_role: int, maintainer_role: int):
     try:
-        update_data = req.model_dump(exclude_none=True)
-        params = jsonable_encoder(update_data)
-        del params['c_id']
-        response = courses_table.update(params).eq('id', req.c_id).execute()
-        return response
-    except Exception as e:
+        course_role_res = (supabase.table("user_courses").select("role_id").
+                       eq('user_id', user_id).
+                       eq("course_id", req.c_id).
+                       execute())
+        course_role = course_role_res.data
+        for role in course_role:
+            if role["role_id"] == admin_role or role["role_id"] == maintainer_role:
+                update_data = req.model_dump(exclude_none=True)
+                params = jsonable_encoder(update_data)
+                del params['c_id']
+                response = courses_table.update(params).eq('id', req.c_id).execute()
+                return response
+        raise NoPermissionException
+    except ValueError as e:
         return None
 
-def delete_course(course_id: int):
+def delete_course(course_id: int, user_id: str, admin_role: int):
     try:
-        response = courses_table.delete().eq('id', course_id).execute()
-        return response
-    except Exception as e:
+        course_role_res = (supabase.table("user_courses").select("role_id").
+                            eq('user_id', user_id).
+                            eq("course_id", course_id).
+                            execute())
+        course_role = course_role_res.data
+        for role in course_role:
+            if role["role_id"] == admin_role:
+                response = courses_table.delete().eq('id', course_id).execute()
+                return response
+        raise NoPermissionException
+    except ValueError as e:
         return None
